+++ mix/load.go
@@ -1,6 +1,6 @@
 package mix
 
-func (c *Computer) lda(aa Word, i, f, op, m int) int {
+func (c *Computer) lda(aa Word, i, f, op, m int) int64 {
 	v := c.Contents[m].Field(f)
 	r := op - LDA
 	if r >= I1 && r <= I6 {
@@ -12,7 +12,7 @@ func (c *Computer) lda(aa Word, i, f, op, m int) int {
 	return 2
 }
 
-func (c *Computer) ldan(aa Word, i, f, op, m int) int {
+func (c *Computer) ldan(aa Word, i, f, op, m int) int64 {
 	v := c.Contents[m].Field(f).Negate()
 	r := op - LDAN
 	if r >= I1 && r <= I6 {
@@ -24,7 +24,7 @@ func (c *Computer) ldan(aa Word, i, f, op, m int) int {
 	return 2
 }
 
-func (c *Computer) sta(aa Word, i, f, op, m int) int {
+func (c *Computer) sta(aa Word, i, f, op, m int) int64 {
 	c.Contents[m].SetField(f, c.Reg[op-STA])
 	return 2
 }
